% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.r
\name{add.nodes}
\alias{add.nodes}
\title{Adding Nodes to DAG}
\usage{
add.nodes(DAG, nodes)
}
\arguments{
\item{DAG}{DAG object}

\item{nodes}{A node or several nodes returned from a call to \code{node} function. If the node(s) under same name(s) already exist, the old node(s) get overwritten.}
}
\value{
An updated DAG object with new nodes
}
\description{
Adding nodes to a growing DAG object, as in \code{DAG + node(...)}. Use either syntax \code{DAG + node} or \code{add.action(DAG, node)}, both give identical results.
}

