% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG_node_R6.R
\docType{class}
\name{DAG_Class}
\alias{DAG_Class}
\title{Class for storing a DAG object (the data generating distribution specified by the SEM)}
\format{An R6 class object.}
\usage{
DAG_Class
}
\details{
Following fields are created during initialization
\itemize{
\item{nodes} ...
\item{subset_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
\item{Kmax} ...
}
Description...
Description...
}
\keyword{datasets}

