% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{igraph.to.sparseAdjMat}
\alias{igraph.to.sparseAdjMat}
\title{Convert igraph Network Object into Sparse Adjacency Matrix}
\usage{
igraph.to.sparseAdjMat(igraph_network)
}
\arguments{
\item{igraph_network}{Network as an \code{igraph} object}
}
\value{
Sparase adjacency matrix returned by \code{igraph::as_adjacency_matrix} function.
NOTE: for directed graphs the friend IDs pointing into vertex \code{i} are assumed to be listed in the column \code{i}
(i.e, \code{which(adjmat[,i])} are friends of \code{i}).
}
\description{
Convert igraph network object into its sparse adjacency matrix representation using \code{as_adjacency_matrix} function from the \code{igraph} package.
}
\seealso{
\code{\link{network}}; \code{\link{sparseAdjMat.to.NetInd}}; \code{\link{NetInd.to.sparseAdjMat}}; \code{\link{sparseAdjMat.to.igraph}};
}

