% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DAG_node_R6.R
\docType{class}
\name{Node_Class}
\alias{Node_Class}
\title{Class for storing a node object (the data generating distribution specified by the SEM)}
\format{An R6 class object.}
\usage{
Node_Class
}
\description{
Class for storing a node object (the data generating distribution specified by the SEM)
}
\details{
Following fields are created during initialization
\itemize{
\item{nodes} ...
\item{subset_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
\item{Kmax} ...
}
Description...
Description...
}
\keyword{datasets}

