% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcausal.r
\name{plotSurvEst}
\alias{plotSurvEst}
\title{(EXPERIMENTAL) Plot Survival}
\usage{
plotSurvEst(surv_res_est = list(), xindx = NULL, y_lab = "", miny = 0,
  maxy = 1)
}
\arguments{
\item{surv_res_est}{A list of vectors containing action-specific discrete survival values over time}

\item{xindx}{A vector of indices for subsetting each of the survival survival vectors in surv_res_est}

\item{y_lab}{A character vector for labeling y axis}

\item{miny}{Minimum y to plot, 0 to 1}

\item{maxy}{Maximum y to plot, 0 to 1}
}
\description{
Plot survival curves given a list of survival outcomes
}

