% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcausal.r
\name{print.DAG.action}
\alias{print.DAG.action}
\title{Print Action Object}
\usage{
\method{print}{DAG.action}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Other arguments to generic print}
}
\description{
Print Action Object
}

