% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcausal.r
\name{print.DAG.node}
\alias{print.DAG.node}
\title{Print DAG.node Object}
\usage{
\method{print}{DAG.node}(x, ...)
}
\arguments{
\item{x}{A Node object}

\item{...}{Other arguments to generic print}
}
\description{
Print DAG.node Object
}

