% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcausal.r
\name{print.DAG}
\alias{print.DAG}
\title{Print DAG Object}
\usage{
\method{print}{DAG}(x, ...)
}
\arguments{
\item{x}{A DAG object.}

\item{...}{Other arguments to generic print.}
}
\description{
Print DAG Object
}

