% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{rconst}
\alias{rconst}
\title{Constant (Degenerate) Node Distribution}
\usage{
rconst(n, const)
}
\arguments{
\item{n}{Sample size.}

\item{const}{Either a vector with one constant value (replicated \code{n} times)
or a vector of length \code{n} or a matrix with \code{n} rows (for a multivariate node).}
}
\value{
A vector of constants of length \code{n}.
}
\description{
Wrapper for constant value (degenerate) distribution.
}
\examples{
#---------------------------------------------------------------------------------------
# Specifying and simulating from a DAG with 1 Bernoulli and 2 constant nodes
#---------------------------------------------------------------------------------------
D <- DAG.empty()
D <- D + node("W1", distr="rbern", prob=0.05)
D <- D + node("W2", distr="rconst", const=1)
D <- D + node("W3", distr="rconst", const=ifelse(W1==1,5,10))
create_mat <- function(W1, W2, W3) cbind(W1,W2,W3)
# Two identical ways of creating a multivariate node (just repeating W1, W2, W3):
D <- D + node(c("W4.1", "W4.2", "W4.3"), distr="rconst", const=cbind(W1, W2, W3))
D <- D + node(c("cW4.1", "cW4.2", "cW4.3"), distr="rconst", const=create_mat(W1, W2, W3))
Dset <- set.DAG(D)
simdat <- sim(Dset, n=200, rndseed=1)
}

