% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_distributions.R
\name{rnet.gnm}
\alias{rnet.gnm}
\title{Call \code{igraph::sample_gnm} to generate random graph object according to the G(n,m) Erdos-Renyi model}
\usage{
rnet.gnm(n, m_pn)
}
\arguments{
\item{n}{Size of the network graph (number of nodes).}

\item{m_pn}{The total number of edges as a fraction of the sample size \code{n}.}
}
\value{
A matrix with n rows, each row lists the indices of friends connected to that particular observation.
}
\description{
Call \code{igraph::sample_gnm} and convert the output to \code{simcausal} network matrix.
The parameter \code{m} of \code{igraph::sample_gnm} is derived from \code{n} and \code{m_pn} as \code{as.integer(m_pn*n)}
}
\seealso{
\code{\link{rnet.gnp}}
}

