% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_distributions.R
\name{rnet.gnp}
\alias{rnet.gnp}
\title{Call \code{igraph::sample_gnp} to generate random graph object according to the G(n,p) Erdos-Renyi model}
\usage{
rnet.gnp(n, p, ...)
}
\arguments{
\item{n}{Size of the network graph (number of nodes).}

\item{p}{Same as \code{igraph::sample_gnp}: The probability for drawing an edge between two arbitrary vertices (G(n,p) graph).}
}
\value{
A matrix with n rows, each row lists the indices of friends connected to that particular observation.
}
\description{
Call \code{igraph::sample_gnp} and convert the output to \code{simcausal} network matrix.
}
\seealso{
\code{\link{rnet.gnm}}
}

