% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{simcausalNEWS}
\alias{simcausalNEWS}
\title{Show the NEWS file for the simcausal package}
\usage{
simcausalNEWS(...)
}
\arguments{
\item{\dots}{additional arguments passed to \code{RShowDoc}}
}
\value{
A invisible path to the NEWS file
}
\description{
Show the NEWS file for the simcausal package
}

