% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcausal.r
\name{vecfun.print}
\alias{vecfun.print}
\title{Print names of custom vectorized functions}
\usage{
vecfun.print()
}
\value{
A vector of vectorized function names
}
\description{
Print current user-defined vectorized function names
}

