% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcausal.r
\name{vecfun.remove}
\alias{vecfun.remove}
\title{Remove Custom Vectorized Functions}
\usage{
vecfun.remove(vecfun_names)
}
\arguments{
\item{vecfun_names}{A character vector of function names that will be removed from the custom list}
}
\value{
An old vector of user-defined vectorized function names
}
\description{
Remove user-defined function names from a global list of custom vectorized functions. See vignette for more details.
}
