% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcausal.r
\name{vecfun.reset}
\alias{vecfun.reset}
\title{Reset Custom Vectorized Function List}
\usage{
vecfun.reset()
}
\value{
An old vector of user-defined vectorized function names
}
\description{
Reset a listing of user-defined vectorized functions.
}

