% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.r
\name{add.nodes}
\alias{add.nodes}
\alias{+.DAG}
\title{Adding Node(s) to DAG}
\usage{
add.nodes(DAG, nodes)

\method{+}{DAG}(obj1, obj2)
}
\arguments{
\item{DAG}{DAG object}

\item{nodes}{A node or several nodes returned from a call to \code{node} function. If the node(s) under same name(s) already exist, the old node(s) get overwritten.}

\item{obj1}{Object that belongs to either classes: \code{DAG}, \code{DAG.action} or \code{DAG.nodelist}.}

\item{obj2}{Object that belongs to either classes: \code{DAG}, \code{DAG.action} or \code{DAG.nodelist}.}
}
\value{
An updated DAG object with new nodes
}
\description{
Adding nodes to a growing DAG object, as in \code{DAG + node()}. Use either syntax \code{DAG + node()} or \code{add.nodes(DAG = , nodes = node())}. Both give identical results, see the examples in the vignette and below for details.
}
\seealso{
\code{\link{node}}
}
