% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parserfunctions_R6.r
\docType{class}
\name{Define_sVar}
\alias{Define_sVar}
\title{Class for defining and evaluating user-specified summary measures (exprs_list)}
\format{An R6 class object.}
\usage{
Define_sVar
}
\description{
Class for defining and evaluating user-specified summary measures (exprs_list)
}
\details{
Following fields are created during initialization
\itemize{
\item{nodes} ...
\item{subset_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
\item{Kmax} ...
}
Evaluates and and stores arbitrary summary measure expressions.
The expressions (exprs_list) are evaluated in the environment of the input data.frame.
}
\keyword{datasets}

