% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.r
\name{DF.to.longDT}
\alias{DF.to.longDT}
\title{Convert \code{data.frame} into long format (faster than \code{DF.to.long})}
\usage{
DF.to.longDT(df_wide)
}
\arguments{
\item{df_wide}{A \code{data.frame} in wide format}
}
\value{
A \code{data.table} object in long format
}
\description{
This utility function takes a simulated data.frame in wide format as an input and converts it into a long format
using \pkg{data.table} package functions \code{melt.data.table} and \code{dcast.data.table}.
}
\details{
Keeps all covariates that appear only once and at the first time-point constant (carry-forward).

All covariates that appear fewer than range(t) times are imputed with NA for missing time-points.

Observations with all NA's for all time-varying covariates are removed.

When removing NA's the time-varying covariates that are attributes (attnames) are not considered.
}

