% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcausal.r
\name{plotSurvEst}
\alias{plotSurvEst}
\title{(EXPERIMENTAL) Plot Discrete Survival Function(s)}
\usage{
plotSurvEst(surv = list(), xindx = NULL, ylab = "", xlab = "t",
  ylim = c(0, 1), ...)
}
\arguments{
\item{surv}{A list of vectors, each containing action-specific discrete survival probabilities over time.}

\item{xindx}{A vector of indices for subsetting the survival vectors in \code{surv}, if omitted all survival probabilities in each \code{surv[[i]]} are plotted.}

\item{ylab}{An optional title for y axis, passed to \code{plot}.}

\item{xlab}{An optional title for x axis, passed to \code{plot}.}

\item{ylim}{Optional y limits for the plot, passed to \code{plot}.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plot discrete survival curves from a list of discrete survival probabilities by calling \code{plot} with \code{type='b'}.
}

