% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.r
\name{simobs}
\alias{simobs}
\title{Simulate Observed Data}
\usage{
simobs(DAG, n, wide = TRUE, LTCF = NULL, rndseed = NULL)
}
\arguments{
\item{DAG}{A DAG objects that has been locked with set.DAG(DAG). Observed data from this DAG will be simulated.}

\item{n}{Number of observations to sample.}

\item{wide}{A logical, if TRUE the output data is generated in wide format, if FALSE, the output longitudinal data in generated in long format}

\item{LTCF}{If forward imputation is desired for the missing variable values, this argument should be set to the name of the node that indicates the end of follow-up event. See the vignette, \code{\link{sim}} and \code{\link{doLTCF}} for additional details.}

\item{rndseed}{Seed for the random number generator.}
}
\value{
A data.frame where each column is sampled from the conditional distribution specified by the corresponding \code{DAG} object node.
}
\description{
This function simulates observed data from a DAG object.
}
\seealso{
\code{\link{doLTCF}} for forward imputing the missing values in already simulating data.

Other simulation.functions: \code{\link{simfull}};
  \code{\link{sim}}
}

